giant initial commit
sometimes these things all have to come together at once, I swear.
It isn't quite perfect or complete and it does disable a couple
things here and there that were rarely used and apparently broken.

On the other hand, it fixes quite a few important things that I
didn't even know were broken and takes a couple giant steps to wards
a brighter future.
